+++ lib/src/visitor/async_evaluate.dart
@@ -1853,6 +1853,18 @@ final class _EvaluateVisitor
   Future<Value?> visitIncludeRule(IncludeRule node) async {
     var mixin = _addExceptionSpan(node,
         () => _environment.getMixin(node.name, namespace: node.namespace));
+    if (node.originalName.startsWith('--') &&
+        mixin is UserDefinedCallable &&
+        !mixin.declaration.originalName.startsWith('--')) {
+      _warn(
+          'Sass @mixin names beginning with -- are deprecated for forward-'
+          'compatibility with plain CSS mixins.\n'
+          '\n'
+          'For details, see https://sass-lang.com/d/css-function-mixin',
+          node.nameSpan,
+          Deprecation.cssFunctionMixin);
+    }
+
     var contentCallable = node.content.andThen((content) => UserDefinedCallable(
         content, _environment.closure(),
         inDependency: _inDependency));
@@ -2508,6 +2520,19 @@ final class _EvaluateVisitor
 
       function = (_stylesheet.plainCss ? null : _builtInFunctions[node.name]) ??
           PlainCssCallable(node.originalName);
+    }
+
+    if (node.originalName.startsWith('--') &&
+        function is UserDefinedCallable &&
+        !function.declaration.originalName.startsWith('--')) {
+      _warn(
+        'Sass @function names beginning with -- are deprecated for forward-'
+        'compatibility with plain CSS functions.\n'
+        '\n'
+        'For details, see https://sass-lang.com/d/css-function-mixin',
+        node.nameSpan,
+        Deprecation.cssFunctionMixin,
+      );
     }
 
     var oldInFunction = _inFunction;