+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 7788c21fd8c721992490ac01d0ef4783dddf3f1f
+// Checksum: 116b8079719577ac6e4dad4aebe403282136e611
 //
 // ignore_for_file: unused_import
 
@@ -1845,6 +1845,18 @@ final class _EvaluateVisitor
   Value? visitIncludeRule(IncludeRule node) {
     var mixin = _addExceptionSpan(node,
         () => _environment.getMixin(node.name, namespace: node.namespace));
+    if (node.originalName.startsWith('--') &&
+        mixin is UserDefinedCallable &&
+        !mixin.declaration.originalName.startsWith('--')) {
+      _warn(
+          'Sass @mixin names beginning with -- are deprecated for forward-'
+          'compatibility with plain CSS mixins.\n'
+          '\n'
+          'For details, see https://sass-lang.com/d/css-function-mixin',
+          node.nameSpan,
+          Deprecation.cssFunctionMixin);
+    }
+
     var contentCallable = node.content.andThen((content) => UserDefinedCallable(
         content, _environment.closure(),
         inDependency: _inDependency));
@@ -2484,6 +2496,19 @@ final class _EvaluateVisitor
 
       function = (_stylesheet.plainCss ? null : _builtInFunctions[node.name]) ??
           PlainCssCallable(node.originalName);
+    }
+
+    if (node.originalName.startsWith('--') &&
+        function is UserDefinedCallable &&
+        !function.declaration.originalName.startsWith('--')) {
+      _warn(
+        'Sass @function names beginning with -- are deprecated for forward-'
+        'compatibility with plain CSS functions.\n'
+        '\n'
+        'For details, see https://sass-lang.com/d/css-function-mixin',
+        node.nameSpan,
+        Deprecation.cssFunctionMixin,
+      );
     }
 
     var oldInFunction = _inFunction;