+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 1.77.2
 
+* Don't emit deprecation warnings for functions and mixins beginning with `__`.
+
+* Allow user-defined functions whose names begin with `_` and otherwise look
+  like vendor-prefixed functions with special CSS syntax.
+
 ### Command-Line Interface
 
 * Properly handle the `--silence-deprecation` flag.